refactor: Remove `LLMProviderAware` interface and inline shared logic
- Deleted `LLMProviderAware` and moved `requireMatchingProvider` to `LLModel`.
- Updated `GoogleGenaiEmbeddingProvider` and `GoogleGenaiLLMClient` to use the new `LLModel.requireMatchingProvider` method directly.
- Simplified provider validation logic across implementations.
- Adjusted affected API signatures and documentation for consistency.
