+++ prompt/prompt-executor/prompt-executor-clients/prompt-executor-google-genai-client/api/prompt-executor-google-genai-client.api
@@ -1,10 +1,10 @@
-public class ai/koog/prompt/executor/clients/google/genai/GoogleGenaiEmbeddingProvider : ai/koog/prompt/executor/clients/LLMEmbeddingProvider, ai/koog/prompt/executor/clients/LLMProviderAware {
+public class ai/koog/prompt/executor/clients/google/genai/GoogleGenaiEmbeddingProvider : ai/koog/prompt/executor/clients/LLMEmbeddingProvider {
 	public fun <init> (Lcom/google/genai/Client;)V
 	public fun <init> (Lcom/google/genai/Client;Lai/koog/prompt/llm/LLMProvider;)V
 	public fun <init> (Lcom/google/genai/Client;Lai/koog/prompt/llm/LLMProvider;Lcom/google/genai/types/EmbedContentConfig;)V
 	public synthetic fun <init> (Lcom/google/genai/Client;Lai/koog/prompt/llm/LLMProvider;Lcom/google/genai/types/EmbedContentConfig;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
 	public fun embed (Ljava/lang/String;Lai/koog/prompt/llm/LLModel;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public fun llmProvider ()Lai/koog/prompt/llm/LLMProvider;
+	public final fun llmProvider ()Lai/koog/prompt/llm/LLMProvider;
 }
 
 public class ai/koog/prompt/executor/clients/google/genai/GoogleGenaiLLMClient : ai/koog/prompt/executor/clients/LLMClient {
