+++ prompt/prompt-executor/prompt-executor-clients/prompt-executor-google-genai-client/src/test/kotlin/ai/koog/prompt/executor/clients/google/genai/GoogleGenaiStreamingTest.kt
@@ -31,6 +31,6 @@ class GoogleGenaiStreamingTest {
                 model = TestModels.noCap
             ).collect {}
         }
-        error.message shouldContain "does not support chat completions"
+        error.message shouldContain "does not support completion capability"
     }
 }
