+++ prompt/prompt-executor/prompt-executor-clients/src/commonMain/kotlin/ai/koog/prompt/executor/clients/LLMClientAPI.kt
@@ -3,6 +3,7 @@ package ai.koog.prompt.executor.clients
 import ai.koog.agents.core.tools.ToolDescriptor
 import ai.koog.prompt.dsl.ModerationResult
 import ai.koog.prompt.dsl.Prompt
+import ai.koog.prompt.llm.LLMProvider
 import ai.koog.prompt.llm.LLModel
 import ai.koog.prompt.message.LLMChoice
 import ai.koog.prompt.message.Message
@@ -14,7 +15,7 @@ import kotlinx.coroutines.flow.Flow
 /**
  * API for [LLMClient]
  */
-public interface LLMClientAPI : LLMProviderAware, AutoCloseable {
+public interface LLMClientAPI : AutoCloseable {
     /**
      * Executes a prompt and returns a list of response messages.
      *
@@ -87,6 +88,13 @@ public interface LLMClientAPI : LLMProviderAware, AutoCloseable {
     public suspend fun models(): List<LLModel> {
         throw UnsupportedOperationException("Not implemented for this client")
     }
+
+    /**
+     * Retrieves the LLMProvider instance associated with this client.
+     *
+     * @return The LLMProvider instance used for executing prompts and managing LLM operations.
+     */
+    public fun llmProvider(): LLMProvider
 
     /**
      * Standard JSON schema generator supported by the LLMClient.
