+++ prompt/prompt-llm/src/commonMain/kotlin/ai/koog/prompt/llm/LLModel.kt
@@ -36,3 +36,16 @@ public fun LLModel.requireCapability(capability: LLMCapability): Unit =
     require(supports(capability)) {
         "Model $id does not support ${capability.id} capability."
     }
+
+/**
+ * Ensures that the provided [LLModel] is associated with the same [LLMProvider].
+ * Throws an [IllegalArgumentException] if the providers do not match.
+ *
+ * @param llmProvider The [LLMProvider] to compare against the current [LLModel].
+ */
+public fun LLModel.requireMatchingProvider(llmProvider: LLMProvider) {
+    require(this.provider == llmProvider) {
+        "Model provider mismatch: ${this.id}.provider=${this.provider}, " +
+            "${this::class.simpleName ?: "(LLMProviderAware)"}.llmProvider=$llmProvider"
+    }
+}
